chore: build web image
