+++ pkg/bot/src/type/cmd.ts
@@ -27,6 +27,7 @@ export const generalEntrySchema = z.object({
   commentRepo: z.string(),
   commentIssueNumber: z.number(),
   commentId: z.number(),
+  namespace: z.string(),
 });
 export type GeneralEntry = z.infer<typeof generalEntrySchema>;
