+++ pkg/def-manager/data/buildspecs/build-container.yml
@@ -14,6 +14,7 @@ phases:
       - echo "IMAGE_TAG=$IMAGE_TAG"
       - echo "GIT_URL=$GIT_URL"
       - echo "GIT_FETCH=$GIT_FETCH"
+      - echo "BUILD_DOCKERFILE=$BUILD_DOCKERFILE"
       - echo ==== info ====
       - pwd
       - whoami
@@ -44,7 +45,7 @@ phases:
   build:
     commands:
       - echo ==== build the Docker image ====
-      - docker build . -f ./docker/api/Dockerfile -t "$IMAGE_REPO_NAME:$IMAGE_TAG"
+      - docker build . -f "$BUILD_DOCKERFILE" -t "$IMAGE_REPO_NAME:$IMAGE_TAG"
       - docker tag $IMAGE_REPO_NAME:$IMAGE_TAG $AWS_ACCOUNT_ID.dkr.ecr.$AWS_DEFAULT_REGION.amazonaws.com/$IMAGE_REPO_NAME:$IMAGE_TAG
       - echo ==== finished ====
   post_build:
