+++ pkg/def-manager/src/stack/operate-env.ts
@@ -103,6 +103,7 @@ export class OperateEnv extends Resource {
 
   readonly computedOpEnv: ComputedOpEnv = {
     API_REPO_NAME: this.parent.apiDevRepo.name,
+    WEB_REPO_NAME: this.parent.webDevRepo.name,
     AWS_ACCOUNT_ID: this.parent.options.sharedEnv.AWS_ACCOUNT_ID,
     S3BACKEND_REGION: this.tfstate.region,
     S3BACKEND_BUCKET: z.string().parse(this.tfstate.bucket),
