+++ pkg/shared/lib/bot-env/index.ts
@@ -12,7 +12,9 @@ export const computedBotEnvSchema = z.object({
   TABLE_NAME: z.string(),
   SSM_PREFIX: z.string(),
   API_REPO_NAME: z.string(),
+  WEB_REPO_NAME: z.string(),
   API_BUILD_PROJECT_NAME: z.string(),
+  WEB_BUILD_PROJECT_NAME: z.string(),
   OPERATE_ENV_PROJECT_NAME: z.string(),
 });
 export type ComputedBotEnv = z.infer<typeof computedBotEnvSchema>;
